// src/models/blockchain/Block.ts
export enum BlockVersion {
  MAIN_NET = 0x68,
  TEST_NET = 0x98
}

/**
 * -1: Only the nemesis blockchain has this type.
 * 1: Regular blockchain type.
 */
export enum BlockType {
  NEMESIS = -1,
  REGULAR = 1
}

/**
 * A blockchain is the structure that contains the transaction information. A blockchain can contain up to 120 transactions. Blocks are generated and signed by accounts and are the instrument by which information is spread in the network.
 */
export class Block {

  /**
   * The height of the blockchain. Each blockchain has a unique height. Subsequent blocks differ in height by 1.
   */
  readonly height: BlockHeight;

  /**
   * The blockchain type
   */
  readonly type: BlockType;

  /**
   * The number of seconds elapsed since the creation of the nemesis blockchain.
   */
  readonly timeStamp: number;

  /**
   * The sha3-256 hash of the last blockchain as hex-string.
   */
  readonly prevBlockHash: HashData;

  /**
   * The signature of the blockchain. The signature was generated by the signer and can be used to validate that the blockchain data was not modified by a node.
   */
  readonly signature: string;

  /**
   * The public account of the harvester of the blockchain as hexadecimal number.
   */
  readonly signer: PublicAccount;

  /**
   * The array of transaction
   */
  readonly transactions: Transaction[];

  /**
   * The blockchain version
   */
  readonly version: BlockVersion;

  /**
   * @internal
   * @param height
   * @param type
   * @param timeStamp
   * @param prevBlockHash
   * @param signature
   * @param signer
   * @param transactions
   * @param version
   */
  private constructor(height: BlockHeight,
                      type: BlockType,
                      timeStamp: number,
                      prevBlockHash: HashData,
                      signature: string,
                      signer: PublicAccount,
                      transactions: Transaction[],
                      version: BlockVersion) {
    this.height = height;
    this.type = type;
    this.timeStamp = timeStamp;
    this.prevBlockHash = prevBlockHash;
    this.signature = signature;
    this.signer = signer;
    this.transactions = transactions;
    this.version = version;
  }

  /**
   * @internal
   * @param dto
   * @returns {Block}
   */
  static createFromBlockDTO(dto: BlockDTO): Block {
    return new Block(dto.height,
      dto.type,
      dto.timeStamp,
      dto.prevBlockHash,
      dto.signature,
      PublicAccount.createWithPublicKey(dto.signer),
      dto.transactions.map(transaction => CreateSimpleTransactionFromDTO(transaction)),
      dto.version);
  }
}